Add interfaces to all pieces of Conversation container
